requirements.txt
click>=8.0
python-dotenv>=1.0
requests>=2.28
